phase5_text_extraction/requirements.txt
google-cloud-vision==3.10.2
opencv-python==4.9.0.80
numpy==1.24.3